+++ week3/week3-assignment/src/components/Modal/Modal.styled.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-const ModalContainer = styled.div`
+export const ModalContainer = styled.div`
   background-color: ${({ theme }) => theme.colors.LightPurple};
   display: flex;
   flex-direction: column;
@@ -17,30 +17,15 @@ const ModalContainer = styled.div`
   gap: 3rem;
 `;
 
-const ModalComment = styled.h1`
+export const ModalComment = styled.h1`
   font-size: 2rem;
   color: white;
 `;
 
-const ModalResetBtn = styled.button`
+export const ModalResetBtn = styled.button`
   background-color: ${({ theme }) => theme.colors.Violet};
   color: white;
   width: 80%;
   height: 5rem;
   border-radius: 1rem;
 `;
-
-export default function Modal({ onClick }) {
-  return (
-    <ModalContainer>
-      <ModalComment>😍축하합니다!😍</ModalComment>
-      <ModalResetBtn
-        onClick={() => {
-          onClick();
-        }}
-      >
-        게임으로 돌아가기
-      </ModalResetBtn>
-    </ModalContainer>
-  );
-}
