--- week3/week3-assignment/src/components/NameChange.jsx
@@ -1,14 +1,0 @@
-import { useState } from "react";
-
-export default function NameChange() {
-    const [name, setName] = useState('thisishwarang')
-    const handleName = () => {
-        setName("coolguy")
-    }
-    return (
-        <div onClick={handleName}>
-            {name}
-        </div>
-    );
-}
-
